Limit Repository Sources for Deployment
This patch limits the source repositories on which the automated
deployment is run to a few trusted sources. The reason for this is that
running code on `pull_request_target` events allows access to security
tokens which can be extracted if arbitrary code is used.

Limiting the sources should fix the problem since the workflow
configuration is always retrieved from the main repository so that no
one can easily change these filters in a pull request from another
source.