Attempt to determine duration if not given
The duration return by the backend can be "null". This is not handled by the frontend at all, resulting in an erroneous state.
Now, the frontend will try to determine the duration from the primary video player if there still is no duration by the time the video player is done loading.