Simple Track Selection (#400)
This closes #398

* Track Selection

This patch adds the ability to select or deselect audio and video
streams of tracks to allow for the removal of unwanted material.

* Simplify User Interface

This patch moves from individual stream selectors, allowing users to
select stream combinations which would fail later on, to a simple
two-button layout to either deactivate a track completely or select an
audio stream as main stream to use for all tracks.

* Accessibility & Internationalization

This patch adds keyboard navigation capabilities, aria labels,
tooltips and loads all texts from the language files.

* Make Track Selection Optional

This patch allows administrators to hide the track selection. This could
potentially avoid problems in some situations, since the feature relies
on some functionality being built into the workflows.

* Additional Explanation

This patch adds some additional information about what happens when
choosing different track selection options.

* Automatically Loose Focus

This patch makes the buttons loose focus after they have been clicked to
avoid styling problems wither enlarged icons.

* Colorful Track States

This tries to make the different track states easier to understand by
changing the button text when you cannot delete a track and adding some
colors to highlight what action is being taken.

* Fix Main Audio Selection

This patch fixes a problem when a single audio track is selected to be
copied over to all other tracks which wasn't properly stored and sent to
Opencast.

* Remove Audio Handling for Improved Usability

The separate handling of audio streams proofed to be confusing and can
easily cause problems since the backend is highly configurable. This
patch removes the separate handling for audio streams for now.