Improved highlighting of segments
Replaces the outline for segments hovered by the scrubber. This should help with visual clarity (the dashed line was really not clear on that front).

Furthermore, adds an outline if the segment is in focus, as is current standard in Chrome and derivatives. This aims to help users relying on keyboard controls.