Added box-shadow outline to workflow selection buttons
This should make it more obvious to users that they are actually buttons, as well as differentiate them from the text above and below them.

Still not quite happy with styling these similar to other buttons, as they are about a "selection", while other buttons are about performing "actions". Would be nice to get that across somehow.