Include Build Version
This patch automatically adds the build version (git commit hash) and
build date to the release so that it is easier for developers and admins
to find out which version is actually deployed. It is only an HTML meta
field and thus completely invisible to non-technical users.

```html
<!DOCTYPE html>
<html lang="en">
  <head>
    <meta name="git-commit" content="3234bb106a4d8ec0cbbb9ff8ffd192625a7a2aa0">
    <meta name="build-date" content="Tue Oct 12 07:39:52 PM CEST 2021">
    <meta …
```

This fixes #463