GitHub Actions Concurrency
If multiple GitHub Actions workflows try to simultaneously deploy to the
same environment, some of them fail with conflicts. This patch that only
a single workflow accessing the same resource runs at a time.

The test deployment will wait for the other workflows to finish. The
demo deployment to editor.opencast.org and to Crowdin will cancel old
workflows since their data will be overwritten by the new workflow
anyway.