Fix Test Deployment
Adding concurrency groups for the test deployment caused some of the
pending jobs to be canceled as GitHub Actions only allows a single
pending workflow in a concurrency group:

> When a concurrent job or workflow is queued, if another job or
> workflow using the same concurrency group in the repository is in
> progress, the queued job or workflow will be pending. Any previously
> pending job or workflow in the concurrency group will be canceled. To
> also cancel any currently running job or workflow in the same
> concurrency group, specify cancel-in-progress: true.

This patch switches to the [Turnstyle](https://github.com/softprops/turnstyle)
action instead, which will make the job wait for previous workflows.
This should have the intended effect.