Relax Playwright tests for the navigation bar
Currently the Playwright tests for the navigation bar expect four specific menu items to be present. However, every menu item besides "Cutting" and "Finish" is configurable to be present or not. Therefore I'd like to propose to only check for the actually required menu items, instead of some arbitrarily defined set.