Allows video player hotkeys in the subtitle editor by changing the way main menu views are rendered
Previously, all main menu views were rendered on startup, and the ones that were not currently needed were simply hidden.
This causes issues with globalHotkeys, were the same hotkey is overloaded and triggers events in views currently not displayed.
This PR changes rendering so that only the views currently active are also being rendered. Which probably makes more sense anyway.

- Also fixes preview mode hotkey not being displayed correctly

- Also breaks the KeyboardControls View!