Avoid calling dispatch hook on every rerender
Some dispatch hooks were placed uncondoditionally, resulting in them being called on every rerender. While that is in and of itself not an issue (redux will easily notice that the repeated dispatches would not change anything and ignores them), it can trigger redux' infinite-loop-prevention hook, causing the website to crash.