change cursor style to text for react-select multi CreatableSelect
The presenter and contributor fields in the metadata tab use a combined
select input control with multi-select enabled. The actual <input> is
only a few pixels wide, so when hovering over the control, it's not
showing a text cursor indicating the possibility to input new text.

This commit sets the cursor on the `valueContainer` component to `text`
and resets it to `default` on the `multiValue` component.

For the list of available components for React-Select see:
https://react-select.com/components

Fixes: #785

Signed-off-by: Daniel Ziegenberg <daniel@ziegenberg.at>