add check for opencast developers team membership for deployments
Currently, the Test & Deploy workflow only runs for selected pull requests.
Adding a bunch of people to the list and keeping it up-to-date could turn
out to be cumbersome. So this adds an additional workflow job and uses
`TheModdingInquisition/actions-team-membership@v1.0` to check the membership
of the person opening the PR.

Signed-off-by: Daniel Ziegenberg <daniel@ziegenberg.at>