Remove broken cleanup workflow
The cleanup workflow does not work at all right now and has never
actually worked. It has a bunch of different issues explained in #847
and how we want to fix this is open for debate.

This patch removes the workflow to prevent the workflow from failing
every time someone merges a pull request. That way, users don't get a
bunch of warnings sent to them.

This fixes #847.