Show all existing subtitles
This patch shows all existing subtitles that are recognized by Opencast
and sent to the editor in the user interface instead of just showing
subtitles which exactly match a specified language flavor.

Which subtitle tracks are sent from Opencast to the front-end is still
configurable in the editor service.

This also does not allow users to create arbitrary subtitle tracks. They
have to already exist.

This fixes #826