Remove duplicated muiStyle
This patch removes duplicated style for the two dropdown menus in the subtitle area.

Unfortunately, I couldn't easily transfer the style to 'cssStyles.tsx' because the theme was then undefined and I couldn't work with the theme variables.
That's why I split it up like the calendar and used the css selectors.

I don't know if it would be worth the effort to create an extra component for these two menus (like for the tooltips) and then use that. But we could get rid of some css selectors.