Remove style-property from FunctionButton component
I removed the style property from the component, because it does not make sense to pass it.
Instead of passing the `basicButtonStyle` everytime, I added it directly to the div where the component is defined.
Without the `basicButtonStyle` the buttons are not properly styled anymore (e.g. the icons are not centered anymore and no hover-effect anymore).