Fix the style of dropdown menus in subtitle view
The background-color of the dropdown menus ("Pick a language" and "Video Flavor" in subtitle view) was accidentally transparent on both high-contrast-modes.

The label is not visible after the dropdown was in focus once. This is only visible when you use a deploy link (like test.editor.opencast.org/841/2022-11-25_07-51-22/) I can't reproduce this on localhost.
Removed some css selectors from index.css that belong to the label but don't seem to have any effect.