Fix Failing Playwright Tests
This patch fixes the problem that the initial runs of Playwright fail on
pull requests all the time. While they eventually succeed on the second
run, the failing tests still leave lots of unnecessary warnings on
GitHub Actions.

The problem seems to be that the React test server can sometimes be
awfully slow when starting up, causing the initial runs to fail with an
empty page.

To prevent this, this pull request actually build the Editor instead
before running the tests on the pre-build project.

This fixes #846