Change subtitle identification from flavor to id
With subtitles becoming first class citizens in Opencast,
their flavor cannot be used as an identifier anymore.
This is problematic for the editor, which relies heavily
on identifying subtitles by subflavor. This PR switches
from subflavor based identification to a track identifier
 based identification.

 The frontend will still try to match subtitle tracks from
 the backend to languages defined in the frontend
 settings. However, it now has to do so by "lang" tag,
 which is not guaranteed to exist. In order to avoid
 creating new subtitle tracks in the backend on every
 save, the frontend settings will therefore require that
 a defined language MUST have a "lang" tag.

 Furthermore, this PR also includes a behavioural
 change in which subtitles are displayed in the
 subtitle select view. Before, only subtitles that were
 specified in the frontend settings were displayed.
With this PR, all subtitles will be displayed, regardless
of whether they are specified in the frontens settings
or not. This change was made with the rationale that
subtitle tags are optional, and subtitle tracks should
therefore be editable in the editor even if no "lang"
tag is specified.