Add type checking to i18n
Adds type checking to our translation library.  This will help
us avoid using translation strings that do not exist.

I've use "TFuncKey" as the type for cases where we pass a
variable string  instead of a literal. This could arguably be
improved upon in the future.

Also some related cleanup.