Fix language resources loading
Currently, some language resources will be loaded after the rendering which leads to missing translation key errors. This commit changes the asynchronous loading of the resources file to synchronous loading.

This is not the most elegant solution, as it delays the rendering. In the best case, the files should still be loaded asynchronously and be integrated in the `initialize` promise of `index.tsx`, for example.

Fixes #766