Make Studio and Editor look alike
Opencast Studio and the Opencast Editor are some of the most prominent
new tools of Opencast. It would be nice if the look and feel of both
would be more similar so that it feels like they belong to the same
service.

This patch is a first step. It introduces a header and logo similar to
Studio.

In the future, we can also use this to give users easy access to
language options and other settings like with Studio. But that is
something for future patches.