Drop unused dependencies
Cleans up the package.json a bit by dropping unused
dependencies.

An interesting side effect of this is that we now use a more
modern DateTimePicker in the Metadata tab. I believe it
provides a vastly better experience than the old one, so I
just happily accepted this. Please do give it a try.