Use shared ESLint config and override some rules
This project has a different code style than the shared config in a few
aspects. That's no problem, that's what overrides are for. Even with
those, we benefit a lot from having a shared base. For some of those
overrides, it might be beneficial to remove those (and fix the warnings)
in the future.