Add callback url for integration (#1084)
After finishing the editing, the user can now navigate back to the calling system using a button. For this, the integrating system can pass a `callbackUrl` url parameter to provide a return link. In the editor settings, the prefixes of the allowed callback urls must be provided to prevent malicious callback urls.