Replace react-hotkeys with react-hotkeys-hook
react-hotkeys did not work properly anymore with the
upgrade to React 18. Furthermore, it has been unmaintained
for several years at this point. Therefore, switching to a
newer, maintained library seems a necessity. This commit
introduces react-hotkeys-hook to take over hotkeys
functionality. To a user everything should work like before
and everything it should look like nothing changed
(except for maybe the order on the overview page, but
that never followed any rhyme or reason in the first place).