Don't show certain thumbnail buttons for single track
The "Generate All" and "Use for all tracks" buttons in the thumbnail
view don't really serve a purpose when there is only one track.
"Generate All" does the same thing as "Generate", and "Use for
all tracks" does straight up nothing. Therefore, we might as well
not display them if there is but a single track.