Undo outer container and instead move scroll container contents
The previous fix for showing the head of scrubber/playhead again
caused a number of issues with the position of the scrubber.
This undoes the previous change, so that the scrubber binds to the
correct parent again, which should undo all the issues.

This also includes a stylistic change in moving the timeline a little
lower inside its scroll container, to make the top of the scrubber show
again.