Enforce eslint no-explicit-any
Adds more typing, but also adds a few eslint exceptions.
I tried to keep the exceptions to types were it makes
sense to keep `any` (i.e. third party libraries returning
it or just cases were we don't care about the contents
of a variable). This may be controversial though.