Dynamic skip threshold
The hard-coded threshold of 3s for skipping to the previous segment
instead of skipping to the start of the segment can be too long for
short videos, where visually, you are far away from the actual cut mark.

This patch fixes the problem by making this a dynamic threshold based on
the video duration with a minimum of 0.5s (guarding against a non-set
video duration) and a maximum of 3s.