Fix initial appearance of thumbnail view
We want the boxes for the thumbnails and the
videos to be roughly the same size, and display
the buttons in the thumbnail boxes on the
right if space allows. This was not happening
when first clicking on the thumbnail view,
wasting space and giving it a generally not so
great look. This changes some max widths
around to fix that.