Update and enforce eslint
Updates eslint to major version 9 and removes other eslint
dependencies that are obsolete.
Furthermore, adds typescript and eslint checks to the build command,
so that warnings and errors must be addressed.
Finally, makes the minimum amount of code changes for this to build.
Currently ignored eslint rules may be enabled by subsequent
commits/pull requests.